Repopulate all object IDs on save
